Change HlsStream to a more generic AsyncReadOnlySource
